start of front end query handling for /products display page
Created system for converting an redux state object to be converted into a query string to be sent to the API using RTK query. This includes filtering null values, and converting the object into the ?key=value&key=value query syntax. Likely better handled by a package on a larger scale, but it's a good start.